Fix pre-commit.ci errors

+++ data_rentgen/dto/pagination.py
@@ -39,4 +39,4 @@ class PaginationDTO(Generic[T]):
 
     @property
     def pages_count(self) -> int:
-        return int(math.ceil(self.total_count / float(self.page_size))) or 1
+        return math.ceil(self.total_count / float(self.page_size)) or 1
